Disable cargo-tarpaulin in CI for now
cargo-tarpaulin passes `-C link-dead-code` to rustc, which presumably
causes it to link all generated code, even if it's unused. This is
needed by cargo-tarpaulin to accurately track coverage. Unfortunately,
this causes linking to fail as no defmt logger is declared, and whatever
part of defmt that relies on having a logger declared isn't eliminated
before linking.

We of course shouldn't declare a logger in this crate as this is a
library, so I've commented out the cargo-tarpaulin job for now. It's not
a great loss, seeing as the crate only has a single test at the time of
writing.

A better solution might be to use some kind of dummy logger when
building tests, just to make linking with `-C link-dead-code` work.